Rearrange so that the built-in migration strags are in private/
