+++ lib/auto-migrations/index.js
@@ -12,11 +12,23 @@
 //  ██║ ╚═╝ ██║██║╚██████╔╝██║  ██║██║  ██║   ██║   ██║╚██████╔╝██║ ╚████║███████║
 //  ╚═╝     ╚═╝╚═╝ ╚═════╝ ╚═╝  ╚═╝╚═╝  ╚═╝   ╚═╝   ╚═╝ ╚═════╝ ╚═╝  ╚═══╝╚══════╝
 //
-// Given a set of models and a datastore, migrate the values based on a strategy.
+
+/**
+ * Module dependencies
+ */
 
 var _ = require('@sailshq/lodash');
-var strategies = require('./strategies');
+var strategies = require('./private/strategies');
 
+
+/**
+ * Given a set of models and a datastore, migrate the values based on a strategy.
+ *
+ * @param  {[type]}   strategy [description]
+ * @param  {[type]}   orm      [description]
+ * @param  {Function} cb       [description]
+ * @return {[type]}            [description]
+ */
 module.exports = function runAutoMigrations(strategy, orm, cb) {
   //  ╦  ╦╔═╗╦  ╦╔╦╗╔═╗╔╦╗╔═╗  ┌─┐┌┬┐┬─┐┌─┐┌┬┐┌─┐┌─┐┬ ┬
   //  ╚╗╔╝╠═╣║  ║ ║║╠═╣ ║ ║╣   └─┐ │ ├┬┘├─┤ │ ├┤ │ ┬└┬┘
