+++ lib/auto-migrations/private/strategies/drop.js
