+++ lib/auto-migrations/private/strategies/index.js
