+++ lib/auto-migrations/private/strategies/safe.js
